+++ src/networkLayerPlus/ndnRouting/model/neighbor/state/NdnRoutingNeighborStateDown.cpp
@@ -10,7 +10,7 @@ void NdnRoutingNeighborStateDown::processEvent(NeighborEventType event) {
         case NeighborEventType::HELLO_RECEIVED: {
             // start inactivity timer
             auto timer = IOC->getTimer();
-            string timerName = "inactivity_timer_" + to_string(neighbor->getBelongingInterface()->getInterfaceID()) + "_" + to_string(neighbor->getRouterID());
+            string timerName = "inactivity_timer_" + to_string(neighbor->getBelongingInterface()->getInterfaceID()) + "_" + to_string((unsigned long long)((unsigned long long)(neighbor->getRouterID())));
             NdnRoutingNeighbor* neighborForCapture = neighbor;
             timer->startTimer(timerName, NDNROUTING_ROUTERDEADINTERVAL * 1000, [neighborForCapture](string) -> bool {
                 neighborForCapture->getBelongingInterface()->getProtocol()->getCrobJobHandler()->neighborInactivityCronJob(neighborForCapture);
